Cliente-Python/requirements.txt
flask
pyodbc
flask-sqlalchemy
flask-wtf
wtforms
request
